+++ packages/core/core/types/errors.ts
@@ -1,0 +1,18 @@
+export {
+  APICallError,
+  EmptyResponseBodyError,
+  InvalidArgumentError,
+  InvalidDataContentError,
+  InvalidPromptError,
+  InvalidResponseDataError,
+  InvalidToolArgumentsError,
+  JSONParseError,
+  LoadAPIKeyError,
+  NoObjectGeneratedError,
+  NoSuchToolError,
+  RetryError,
+  ToolCallParseError,
+  TypeValidationError,
+  UnsupportedFunctionalityError,
+  UnsupportedJSONSchemaError,
+} from '@ai-sdk/provider';