+++ packages/mistral/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @ai-sdk/mistral
 
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies [41d5736]
+  - @ai-sdk/provider@0.0.3
+  - @ai-sdk/provider-utils@0.0.5
+
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies [56ef84a]
+  - @ai-sdk/provider-utils@0.0.4
+
 ## 0.0.6
 
 ### Patch Changes